polymarket-compounder/requirements.txt
py-clob-client>=0.5.0
aiohttp>=3.9.0
web3>=6.0.0
python-dotenv>=1.0.0
